pyproject.toml
[tool.poetry]
name = "sam-assistant"
version = "0.1.0"
description = ""
authors = ["Fluttensor <mohammadhosine87@yahoo.com>"]
readme = "README.md"
packages = [{include = "sam_assistant"}]

[tool.poetry.scripts]
sam = "sam.main:app"
gpt4free = "gpt4free.test:main"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.180"
openai = "^0.27.4"
unstructured = "^0.6.1"
faiss-cpu = "^1.7.3"
fastapi = "^0.95.1"
llama-cpp-python = "^0.1.38"
asyncio = "^3.4.3"
sse-starlette = "^1.3.4"
python-dotenv = "^1.0.0"
banana-dev = "^4.0.2"
cohere = "^4.2.0"
huggingface-hub = "^0.13.4"
websocket-client = "^1.5.1"
requests = "^2.29.0"
tls-client = "^0.2"
PyPasser = "^0.0.5"
names = "^0.3.0"
colorama = "^0.4.6"
curl-cffi = "^0.5.5"
streamlit = "1.21.0"
selenium = "^4.9.0"
fake-useragent = "^1.1.3"
chromadb = "^0.3.21"
wikipedia = "^1.4.0"
duckduckgo-search = "^2.9.1"
twocaptcha = "^0.0.1"
pydantic = "^1.10.7"
pyowm = "^3.3.0"
youtube-search = "^2.1.2"
gradio-tools = "^0.0.9"
pdfminer = "^20191125"
pypdf = "^3.9.0"
nest-asyncio = "^1.5.6"
pysrt = "^1.1.2"
playwright = "^1.34.0"
youtube-transcript-api = "^0.6.0"
pytube = "^15.0.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
